+++ src/SanchoBBDO/Codes/Command/AbstractDumpToFileCommand.php
@@ -6,8 +6,23 @@ use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
+/**
+ * @deprecated AbstractDumpToFileCommand is deprectated since v1.1.0 and will be
+ * removed on v1.2.0. Use AbstractCodesCommand instead.
+ */
 abstract class AbstractDumpToFileCommand extends AbstractDumpCommand
 {
+    public function __construct($name = null)
+    {
+        trigger_error(
+            "AbstractDumpToFileCommand is deprectated since v1.1.0 and will " .
+            "be removed on v1.2.0. Use AbstractCodesCommand instead.",
+            E_USER_DEPRECATED
+        );
+
+        parent::__construct($name);
+    }
+
     function init()
     {
         $this