+++ src/SanchoBBDO/Codes/Command/DumpCsvCommand.php
@@ -7,8 +7,23 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
+/**
+ * @deprecated DumpCsvCommand is deprectated since v1.1.0 and will be removed on
+ * v1.2.0. Use DumpCodesCommand instead.
+ */
 class DumpCsvCommand extends AbstractDumpToFileCommand
 {
+    public function __construct($name = null)
+    {
+        trigger_error(
+            "DumpCsvCommand is deprectated since v1.1.0 and will be removed " .
+            "on v1.2.0. Use DumpCodesCommand instead.",
+            E_USER_DEPRECATED
+        );
+
+        parent::__construct($name);
+    }
+
     public function init()
     {
         parent::init();